improved error handling on run_file wrapper

+++ run_file.py
@@ -10,9 +10,15 @@ def main(argv: list[str]):
     source = argv[1]
     target = os.path.join(os.path.dirname(source), 'compiled', os.path.basename(source).replace('.', '_') + '.zo')
     output = run(['c:\\\\program files\\racket\\raco.exe', 'make', source], capture_output=True)
-    output = run(['c:\\\\program files\\racket\\racket.exe', target], capture_output=True).stdout.decode()
+    if output.stderr:
+        print(output.stderr.decode(), end='')
+        return 1
+    output = run(['c:\\\\program files\\racket\\racket.exe', target], capture_output=True)
     shutil.rmtree(os.path.join(os.path.dirname(source), 'compiled'))
-    print(output, end='')
+    if output.stderr:
+        print(output.stderr.decode(), end='')
+        return 1
+    print(output.stdout.decode(), end='')
     return 0
 
 if __name__ == '__main__':
